codegen: rename an internal module
